ecosight-backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.12
motor==3.6.0
pymongo==4.10.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1
ultralytics==8.3.50
opencv-python-headless==4.10.0.84
numpy==1.26.4
Pillow==11.0.0
httpx==0.27.2
websockets==13.1
deep-sort-realtime==1.3.2
scipy==1.14.1
onnx==1.17.0
onnxruntime==1.20.1
aiofiles==24.1.0
python-jose[cryptography]==3.3.0
requests==2.32.3
